start roles and permissions

+++ back-end/resources/views/users/index.blade.php
@@ -4,16 +4,22 @@
             <div class="pull-left">
                 <h2>Users Management</h2>
             </div>
+            @can('user-create')
             <div class="pull-right">
                 <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
             </div>
+                @endcan
         </div>
     </div>
+
+
     @if ($message = Session::get('success'))
         <div class="alert alert-success">
             <p>{{ $message }}</p>
         </div>
     @endif
+
+
     <table class="table table-bordered">
         <tr>
             <th>No</th>
@@ -30,18 +36,25 @@
                 <td>
                     @if(!empty($user->getRoleNames()))
                         @foreach($user->getRoleNames() as $v)
-                            <span class="badge rounded-pill bg-dark">{{ $v }}</span>
+                            <label class="badge badge-success">{{ $v }}</label>
                         @endforeach
                     @endif
                 </td>
                 <td>
                     <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-                    <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
-                    {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
+                   @can('user-edit') <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>@endcan
+                   @can('user-delete')
+                        {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                     {!! Form::close() !!}
+                    @endcan
                 </td>
             </tr>
         @endforeach
     </table>
+
+
     {!! $data->render() !!}
+
+
+    <p class="text-center text-primary"><small>Tutorial by ItSolutionStuff.com</small></p>
